Add sliders, comments
Got it working with two additional sliders: one to change the size of individual agents ("Size") and one to change their tail length ("Relative Tail Length"). Both are straightforward: size simply moved the "agentsize" fixed parameter into the same object format as the other tunable parameters and updates the agents by calling "tadpole" in the "updateCartoon" function calls. "Relative Tail Length" is similar in that it is also invoked in the updated tadpole() calls, but it's a new parameter that replaces a literal in-line constant value applied to the power that is the exponential value on the sine term in the tadpole path generation function.

Also changed color scheme, and working on getting opacity to scale with size. I inverted the shape of the tadpole so that I could change how the transform step updates in order to make sure I'm understanding how that part works.